Allow CVMFS-based group mapping to be done on either the TW or the schedd.
Fixes a few bugs from the prior commit.
